+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## unreleased
 * Add subscription charged unsuccessfully sample webhook to webhook testing gateway
+* Add initial support for GraphQL API's
+* Add Transaction-Level Fee Report API
 
 ## 2.81.0
 * Allow payee ID to be passed in options params for transaction create